Convert decimals to floats

+++ src/digit_recognition/MNISTImage.py
@@ -1,7 +1,4 @@
-from decimal import Decimal
-
-
 class MNISTImage():
-    def __init__(self, image_array: list[Decimal], label: int):
+    def __init__(self, image_array: list[float], label: int):
         self.image_array = image_array
         self.label = label
