+++ tests/neuralnet/test_sigmoid_node_simple_example.py
@@ -1,4 +1,4 @@
-from decimal import Decimal
+
 import math
 import unittest
 
@@ -10,11 +10,11 @@ class TestSigmoidNodeSimpleExample(unittest.TestCase):
 
     def test_two_input_nodes(self):
         node = SigmoidNode()
-        node.bias = Decimal(2)
+        node.bias = float(2)
         
         node.input_synapses = [
-            self.createSynapse(activation=Decimal(1), weight=Decimal(3)),
-            self.createSynapse(activation=Decimal(0.5), weight=Decimal(2))
+            self.createSynapse(activation=float(1), weight=float(3)),
+            self.createSynapse(activation=float(0.5), weight=float(2))
         ]
         
         node.determine_activation()
@@ -25,14 +25,14 @@ class TestSigmoidNodeSimpleExample(unittest.TestCase):
 
 
     def get_expected_activation(self):
-        netInput = (Decimal(1)*Decimal(3)) + (Decimal(0.5)*Decimal(2))
+        netInput = (float(1)*float(3)) + (float(0.5)*float(2))
 
         netInputWithBias = netInput + 2
         
         return self.sigmoid(netInputWithBias)
     
     
-    def createSynapse(self, activation: Decimal, weight: Decimal):
+    def createSynapse(self, activation: float, weight: float):
         input_node = SigmoidNode()
         input_node.activation = activation
         output_node = SigmoidNode()
@@ -42,5 +42,5 @@ class TestSigmoidNodeSimpleExample(unittest.TestCase):
         return synapse
 
 
-    def sigmoid(self, net_input: Decimal):
+    def sigmoid(self, net_input: float):
         return 1 / (1 + math.exp(-net_input))
