+++ tests/neuralnet/weather_predictor/weather_conditions.py
@@ -1,8 +1,8 @@
-from decimal import Decimal
+
 
 
 class WeatherConditions():
     
-    def __init__(self, temperature: Decimal, cloud_cover: Decimal):
+    def __init__(self, temperature: float, cloud_cover: float):
         self.temperature = temperature
         self.cloud_cover = cloud_cover
