+++ tests/test_networks/complex_network_creator.py
@@ -1,4 +1,4 @@
-from decimal import Decimal
+
 from src.neuralnet.network import Network
 
 
@@ -21,40 +21,40 @@ class ComplexNetworkCreator():
 
     
     def set_node_layer_1(self, network: Network):
-        network.node_layers[0][0].starting_input = Decimal('1.5')
-        network.node_layers[0][0].bias = Decimal('2')
+        network.node_layers[0][0].starting_input = float('1.5')
+        network.node_layers[0][0].bias = float('2')
 
-        network.node_layers[0][1].starting_input = Decimal('1.1')
-        network.node_layers[0][1].bias = Decimal('3')
+        network.node_layers[0][1].starting_input = float('1.1')
+        network.node_layers[0][1].bias = float('3')
     
         
     def set_synapse_layer_1(self, network: Network):
-        network.get_synapse_between(network.node_layers[0][0].id, network.node_layers[1][0].id).weight = Decimal('0.1')
-        network.get_synapse_between(network.node_layers[0][0].id, network.node_layers[1][1].id).weight = Decimal('0.4')
-        network.get_synapse_between(network.node_layers[0][0].id, network.node_layers[1][2].id).weight = Decimal('0.5')
+        network.get_synapse_between(network.node_layers[0][0].id, network.node_layers[1][0].id).weight = float('0.1')
+        network.get_synapse_between(network.node_layers[0][0].id, network.node_layers[1][1].id).weight = float('0.4')
+        network.get_synapse_between(network.node_layers[0][0].id, network.node_layers[1][2].id).weight = float('0.5')
         
-        network.get_synapse_between(network.node_layers[0][1].id, network.node_layers[1][0].id).weight = Decimal('0.2')
-        network.get_synapse_between(network.node_layers[0][1].id, network.node_layers[1][1].id).weight = Decimal('0.6')
-        network.get_synapse_between(network.node_layers[0][1].id, network.node_layers[1][2].id).weight = Decimal('1.2')
+        network.get_synapse_between(network.node_layers[0][1].id, network.node_layers[1][0].id).weight = float('0.2')
+        network.get_synapse_between(network.node_layers[0][1].id, network.node_layers[1][1].id).weight = float('0.6')
+        network.get_synapse_between(network.node_layers[0][1].id, network.node_layers[1][2].id).weight = float('1.2')
     
     
     def set_node_layer_2(self, network: Network):
-        network.node_layers[1][0].bias = Decimal('3')
-        network.node_layers[1][1].bias = Decimal('1')
-        network.node_layers[1][2].bias = Decimal('1.1')
+        network.node_layers[1][0].bias = float('3')
+        network.node_layers[1][1].bias = float('1')
+        network.node_layers[1][2].bias = float('1.1')
         
         
     def set_synapse_layer_2(self, network: Network):
-        network.get_synapse_between(network.node_layers[1][0].id, network.node_layers[2][0].id).weight = Decimal('1.2')
-        network.get_synapse_between(network.node_layers[1][0].id, network.node_layers[2][1].id).weight = Decimal('1.4')
+        network.get_synapse_between(network.node_layers[1][0].id, network.node_layers[2][0].id).weight = float('1.2')
+        network.get_synapse_between(network.node_layers[1][0].id, network.node_layers[2][1].id).weight = float('1.4')
         
-        network.get_synapse_between(network.node_layers[1][1].id, network.node_layers[2][0].id).weight = Decimal('0.2')
-        network.get_synapse_between(network.node_layers[1][1].id, network.node_layers[2][1].id).weight = Decimal('0.4')
+        network.get_synapse_between(network.node_layers[1][1].id, network.node_layers[2][0].id).weight = float('0.2')
+        network.get_synapse_between(network.node_layers[1][1].id, network.node_layers[2][1].id).weight = float('0.4')
         
-        network.get_synapse_between(network.node_layers[1][2].id, network.node_layers[2][0].id).weight = Decimal('1.5')
-        network.get_synapse_between(network.node_layers[1][2].id, network.node_layers[2][1].id).weight = Decimal('1.3')
+        network.get_synapse_between(network.node_layers[1][2].id, network.node_layers[2][0].id).weight = float('1.5')
+        network.get_synapse_between(network.node_layers[1][2].id, network.node_layers[2][1].id).weight = float('1.3')
         
         
     def set_node_layer_3(self, network: Network):
-        network.node_layers[2][0].bias = Decimal('0.4')
-        network.node_layers[2][1].bias = Decimal('0.6')
+        network.node_layers[2][0].bias = float('0.4')
+        network.node_layers[2][1].bias = float('0.6')
